services/portal-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.14.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-dotenv==1.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
requests==2.31.0
shapely==2.0.2
geoalchemy2==0.14.2
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
lightgbm==4.1.0
openai>=1.3.0
sentry-sdk[fastapi]==1.39.1
apscheduler>=3.10.0
redis>=5.0.0
